Separated drone entity and physics
