+++ src/main/java/bluevista/fpvracing/client/input/AxisValues.java
@@ -1,9 +1,5 @@
 package bluevista.fpvracing.client.input;
 
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-import net.minecraft.util.math.MathHelper;
-
 public class AxisValues {
     public float currT;
     public float currX;
@@ -22,10 +18,6 @@ public class AxisValues {
         this.set(currT, currX, currY, currZ);
     }
 
-    public AxisValues(AxisValues start, AxisValues end, float delta) {
-        this.set(lerp(start, end, delta));
-    }
-
     public void set(AxisValues axisValues) {
         this.set(axisValues.currT, axisValues.currX, axisValues.currY, axisValues.currZ);
     }
@@ -39,15 +31,5 @@ public class AxisValues {
 
     public String toString() {
         return "[CurrT: " + currT + " CurrX: " + currX + " CurrY: " + currY + " CurrZ: " + currZ + "]";
-    }
-
-    @Environment(EnvType.CLIENT)
-    public static AxisValues lerp(AxisValues start, AxisValues end, float delta) {
-        AxisValues av = new AxisValues();
-        av.currT = MathHelper.lerp(start.currT, end.currT, delta);
-        av.currX = MathHelper.lerp(start.currX, end.currX, delta);
-        av.currY = MathHelper.lerp(start.currY, end.currY, delta);
-        av.currZ = MathHelper.lerp(start.currZ, end.currZ, delta);
-        return av;
     }
 }
