+++ src/main/java/bluevista/fpvracing/client/renderers/DroneRenderer.java
@@ -33,11 +33,11 @@ public class DroneRenderer extends EntityRenderer<DroneEntity> {
         this.shadowRadius = 0.2F;
 
         matrixStack.push();
-        if (droneEntity.isActive()) {
-            matrixStack.peek().getModel().multiply(QuaternionHelper.quat4fToQuaternion(droneEntity.getOrientation()));
+        if (droneEntity.physics.isActive()) {
+            matrixStack.peek().getModel().multiply(QuaternionHelper.quat4fToQuaternion(droneEntity.physics.getOrientation()));
         } else {
             matrixStack.peek().getModel().multiply(QuaternionHelper.quat4fToQuaternion(
-                    QuaternionHelper.slerp(droneEntity.getPrevOrientation(), droneEntity.getOrientation(), delta)
+                    QuaternionHelper.slerp(droneEntity.physics.getPrevOrientation(), droneEntity.physics.getOrientation(), delta)
             ));
         }
