+++ src/main/java/bluevista/fpvracing/mixin/ClientPlayNetworkHandlerMixin.java
@@ -98,8 +98,7 @@ public class ClientPlayNetworkHandlerMixin {
 
         if (type == ServerInitializer.DRONE_ENTITY) {
             float yaw = (float)(packet.getYaw() * 360) / 256.0F;
-            entity = new DroneEntity(world, 0, new Vec3d(x, y, z), yaw);
-            world.spawnEntity(entity);
+            entity = new DroneEntity(world, new Vec3d(x, y, z), yaw);
         }
 
         if (entity != null) {
